+++ packages/renderer/src/lib/preferences/PreferencesResourcesRendering.svelte
@@ -29,7 +29,7 @@ import { context } from '/@/stores/context';
 import { onboardingList } from '/@/stores/onboarding';
 import { providerInfos } from '/@/stores/providers';
 
-import { extractConnectionResourceMetrics, RESOURCE_FORMATS } from './ConnectionResourceMetrics';
+import { extractConnectionResourceMetrics, RESOURCE_FORMATS } from './connection-resource-metrics';
 import { eventCollect } from './preferences-connection-rendering-task';
 import PreferencesConnectionActions from './PreferencesConnectionActions.svelte';
 import PreferencesConnectionsEmptyRendering from './PreferencesConnectionsEmptyRendering.svelte';
