+++ packages/renderer/src/lib/preferences/connection-resource-metrics.ts
